requirements.txt
fastapi==0.95.2
alembic==1.11.1
psycopg2-binary==2.9.6
pydantic==1.10.7
python-dotenv==1.0.0
SQLAlchemy==2.0.14
uvicorn==0.22.0
